Path following of UAV under time varying wind
an adaptive control scheme for
Unmanned Aerial Vehicles (UAVs) path following under slowly
time-varying wind is developed. The proposed control strategy
integrates the path following law based on the vector field
method with an adaptive term counteracting the effect of wind’s
unknown component. The control strategy is developed and
numerically simulated for straight line and orbit following
scenarios. In particular, it is shown that the path following
error is bounded under slowly time-varying unknown wind
and converges to zero for unknown constant wind. Numerical
simulations illustrate that, in environments with unknown and
slowly time-varying wind conditions, the proposed method
compensates for the lack of knowledge of the wind vector,
and attains a smaller path following error than state-of-theart vector
field method.